requirements.txt
openai-whisper
sounddevice
numpy
webrtcvad
scipy
requests
pynput
python-dotenv
elevenlabs
